Label Option and Redeemer in their own files
