+++ src/Cooked/Skeleton/Label.hs
@@ -1,0 +1,36 @@
+-- | Labels are arbitrary information that can be added to skeleton. They are
+-- meant to be pretty-printed. The common use case we currently have is to tag
+-- skeletons that have been modified by tweaks and automated attacks.
+module Cooked.Skeleton.Label
+  ( LabelConstrs,
+    TxLabel (..),
+  )
+where
+
+import Cooked.Pretty.Class
+import Type.Reflection
+
+type LabelConstrs x = (PrettyCooked x, Show x, Typeable x, Eq x, Ord x)
+
+data TxLabel where
+  TxLabel :: (LabelConstrs x) => x -> TxLabel
+
+instance Eq TxLabel where
+  a == x = compare a x == EQ
+
+instance Show TxLabel where
+  show (TxLabel x) = show x
+
+instance PrettyCooked TxLabel where
+  prettyCookedOpt opts (TxLabel x) = prettyCookedOpt opts x
+
+instance Ord TxLabel where
+  compare (TxLabel a) (TxLabel x) =
+    case compare (SomeTypeRep (typeOf a)) (SomeTypeRep (typeOf x)) of
+      LT -> LT
+      GT -> GT
+      EQ -> case typeOf a `eqTypeRep` typeOf x of
+        Just HRefl -> compare a x
+        -- This can never happen, since 'eqTypeRep' is implemented in terms of
+        -- '==' on the type representation:
+        Nothing -> error "Type representations compare as EQ, but are not eqTypeRep"
